Document getting started with UTF-8 & APL symbols
It's an eye-opening experience to jump into a language whose entire
array of characters cannot (immediately) be physically represented by
the most familiar computing interface we know: the keyboard.

And even then, after perusing

    /usr/X11R6/share/X11/locale/en_US.UTF-8/Compose

And comparing it to the Unicomp APL keyboard layout, some symbols are
absent from the file and some keys on the keyboard aren't in the Compose
file.

It was a useful exercise, since patterns emerged from all the sorts of
ways the special characters could combine--and there are not too many of
them to mix, as it were. A combination of composing and toggle should
get us all the way there.

On a webpage, it's different of course. But at least we'll be able to
type the characters themselves.

Changes document getting the ball rolling with typing alpha in OpenBSD.
This is just the beginning.

References: xmodmap(1), setxkbmap(1), xev(1), xkbcomp(1),
xkeyboard-config(7), Unix/Linux Stack Exchange q/151966, a/151982,
a/470135, a/471082, a/228761, DaemonForums 5371, OpenBSD FAQ 10#locales

(#5)