Update docs: features, typos, caveats, more notes
Reviewing the README, found several things to change. First, since we're
going all-in with Unicode, there's not going to be--or possibly even a
real feature in APL--for ASCII-equivalent symbols. I think because even
something like `/\` to mean (logical AND) "up caret" is going to be hard
Fundamentally, human vision sees two characters.

We also want to note that Left Control + Left Alt means no more easily
switching to a console (F2 or thereabouts) for root operations. This is
minor friction, and welcome, since it's bad practice anyway.

We also note Right Alt + Caps is an ISO_Level_3_Shift, which int'l folks
may understand. For us, it's an escape valve to avoid locking ourselves
in, in case we need to map even more mystery characters. This early on,
with very little knowledge, it's all about options, options, options.

Noted to press backtick (or grave) instead of tilde, since asciitilde is
really Shift + backtick, and that could be confusing: we don't press
Shift at all; we just press the backtick key by itself.

Mentioned the underbar in patterns, so we ended up (happily) with a kind
of ascending chorus: underscore, asciitilde, diaresis, and macron.

(#7)