Map initial APL symbols to standard keyboard keys
We add the start of some APL mappings to the keys

    ` ~, 1 !, 2 @, 3 #, 4 $, 5 %, 6 ^, and 7 &

For example, the first key (grave) and its shift (tilde) generate the
APL symbols diamond and quad diamond, respectively, with the Ctrl-Alt
toggle.

One convention we use is to use uppercase letters in the Unicode
code point when it is from /usr/X11R6/include/X11/keysymdef.h file;
otherwise, it is from the /usr/X11R6/share/X11/en_US.UTF-8/Compose file

With OpenBSD base vi, Unicode symbols are output as (so far) two or
three hex codes. That's a useful data point--or could be in the future-
-so we include a table for it. Otherwise, we `cat` the file to stdout
in xterm to see the hex codes "rendered" to the actual single symbol.

Sometimes vi does not print hex codes but an ASCII version of the
symbol. These are notable in that, even though a Unicode code point is
defined in the .Xmodmapapl file, the editor prints straight through of
the symbol (< or >, for example) instead of the _Unicode_ code point.

One symbol we could not find is DEL OVERBAR equivalent, or an
upside-down triangle (DEL) with macron above it. We leave that as a ?
question mark symbol for now.

Between editing the .Xmodmapapl file, one does need to exit out of X and
go back in with `startx` to refresh mappings. The Restart Fvwm from the
context menu didn't seem to reload the file. But when things aren't
working, exiting out lets us see if there were any issues loading the
dot file itself.

Using ed(1) with only these primitives

    x,yp  i|.|w  x  d  -x

is good practice. At least to get started on the other things. It's like
learning vi (again)! Or something.

(#11)