Use Unicode variants where available; CIRCLE MINUS
We've been using the Unicode APL keyboard layout for where symbols
should go. CIRCLE MINUS is now used where we could not find CIRCLE (or
zero?) with a horizontal line through it. They really seem like distinct
symbols. Anyway, we'll go with that.

Used Wikipedia's article on digital encoding of APL symbols for
additional Unicode U+____ codes for previous ASCII symbols like =, *, ',
and ?; it turns out most of them output as the ASCII equivalent anyway:
to vi, hex codes aren't generated.

The only different one was P key, so we have a neat star to go with the
star diaeresis instead of plain ol' asterisk.

Using ed(1) right now is mostly stream of consciousness, so apologies if
it starts to ramble. This is the early phase where we use ed(1) and then
later try to refine our abilities (editing). (Course, could also edit in
git-land elsewhere, but that's cheating.)

(Oh and tilde ~ also. I believe also translates to the ASCII version.)

The hope was Unicode variants are larger or more pronounced, which would
help with the default font size. But, we'll probably have to figure that
out anyway down the road.

So there's a couple ideas for the website now. We're using lynx, so we
can't really test Javascript. Not sure if forms can accumulate input to
itself when the page refreshes. Will have to see.

And we could also just have informational pages too, or comparison pages
whatever would help jog learning and association. I mean, goodness--we
haven't even installed APL yet (!)

There's alternate layouts out there, like

https://web.archive.org/web/20120310113525/https://www.wickensonline.co.uk/apl/union-large.png

I want to assume APL programmers end up customizing their keyboards,
just as vim/neovim & Emacs folks with plugins, dotfiles, and the like.

What's left is to see if we can compose the rest of the symbols, but if
they don't exist in Unicode, well... we'll have to see. But we do have
the Compose key, so we can certainly map everything pretty easily.

Closes #15.